Post request
Created post() to be able to respond when the form is submitted. In html, added action and method attributes to the form element and name attribute to the input element. Action controls which url this form sends its request to, when the user submits the form. Method specifies the http method to be used when submitting form data.